Add files via upload
LED TreatLife product. 
I am just getting started and have finished for tonight. 
I have your controller and node i am going over its polling to see if IP address match up to devices I have, three LED's and two switches for testing and one LED strip.
Now i used the snapshot.json file to read and it pulls all nodes in on my poly then passed everything to the nodes. 
I see with your approach it adds the node then finds it ID, IP and KEY in the node. 
I am adding and passing it from the controller currently. 
So i need to wrap my head around it and send modify the add node to add LED's and switches and we need to ad an identifier and I think the DPS numbers will work. Now LED's use 20 for on off and switches use 1 looks like yours uses 101. So i worked this all out and wanted to run wizard somehow to produce the snapshot and make it readable in the controller.
So here is the LED so far and you can see what i am passing with ADD NODE.